+++ src/main/java/com/ana29/deliverymanagement/global/exception/GlobalExceptionHandler.java
@@ -5,6 +5,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -21,6 +22,29 @@ public class GlobalExceptionHandler {
     @ResponseBody
     public ResponseEntity<RestApiException> illegalArgumentExceptionHandler(IllegalArgumentException ex, HttpServletRequest request) {
         // 회원가입 URL에서 발생한 예외인 경우
+        if (request.getRequestURI().contains("/api/users/**")) {
+            HttpHeaders headers = new HttpHeaders();
+            // 에러 메시지를 URL 인코딩하여 쿼리 파라미터로 전달 (예: ?error=메시지)
+//            인코딩 없을 시 url을 해석하지 못하여 403 에러
+            String encodedErrorMessage = URLEncoder.encode(ex.getMessage(), StandardCharsets.UTF_8);
+            String redirectUrl = "/api/users/sign-up?error=" + encodedErrorMessage;
+            // 302 FOUND 상태와 함께 리다이렉션 헤더를 반환
+            return new ResponseEntity<>(headers, HttpStatus.FOUND);
+        }
+
+        RestApiException restApiException = new RestApiException(ex.getMessage(), HttpStatus.BAD_REQUEST.value());
+        return new ResponseEntity<>(
+                // HTTP body
+                restApiException,
+                // HTTP status code
+                HttpStatus.BAD_REQUEST
+        );
+    }
+
+    @ExceptionHandler({MethodArgumentNotValidException.class})
+    @ResponseBody
+    public ResponseEntity<RestApiException> methodArgumentNotValidExceptionHandler(MethodArgumentNotValidException ex, HttpServletRequest request) {
+        // MethodArgumentNotValidException 발생한 예외인 경우
         if (request.getRequestURI().contains("/api/users/sign-up")) {
             HttpHeaders headers = new HttpHeaders();
             // 에러 메시지를 URL 인코딩하여 쿼리 파라미터로 전달 (예: ?error=메시지)
