+++ src/main/java/com/ana29/deliverymanagement/restaurant/controller/CategoryController.java
@@ -1,6 +1,6 @@
 package com.ana29.deliverymanagement.restaurant.controller;
 
-import com.ana29.deliverymanagement.security.constant.user.UserRoleEnum;
+import com.ana29.deliverymanagement.user.constant.user.UserRoleEnum;
 import com.ana29.deliverymanagement.restaurant.dto.CategoryRequestDto;
 import com.ana29.deliverymanagement.restaurant.dto.CategoryResponseDto;
 import com.ana29.deliverymanagement.security.UserDetailsImpl;
@@ -8,7 +8,6 @@ import com.ana29.deliverymanagement.restaurant.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
