+++ src/main/java/com/ana29/deliverymanagement/security/UserDetailsImpl.java
@@ -1,7 +1,7 @@
 package com.ana29.deliverymanagement.security;
 
 import com.ana29.deliverymanagement.user.entity.User;
-import com.ana29.deliverymanagement.security.constant.user.UserRoleEnum;
+import com.ana29.deliverymanagement.user.constant.user.UserRoleEnum;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
