+++ src/main/java/com/ana29/deliverymanagement/user/dto/SigninRequestDto.java
@@ -9,7 +9,8 @@ import lombok.Setter;
 @Getter
 public class SigninRequestDto {
     @NotBlank
-    @Pattern(regexp = "^(?=.*[a-z])(?=.*[0-9])[a-z0-9^\\s]{4,10}$", message = "닉네임은 4자 이상 10자 이하로 입력해주세요.")
+    @Pattern(regexp = "^(?=.*[a-z])(?=.*[0-9])[a-z0-9^\\s]{4,10}$",
+            message = "닉네임은 알파벳 소문자와 숫자로 이루어진 4자 이상 10자 이하로 입력해주세요.")
     private String id;
     @NotBlank
     @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^A-Za-z\\d\\s])[^\\s]{8,15}$",
